moved ScoreSet to own module
